implementa Iris{Particle, Population, Agent}

+++ src/Model/Impl/Iris/IrisFunction.java
@@ -4,10 +4,10 @@ import Model.Abstraction.Function;
 import Model.Abstraction.Particle;
 import Model.Impl.CoreFactoryCreator;
 
-public class IrisFunction implements Function<int[]> {
+public class IrisFunction implements Function<double[]> {
 
     @Override
-    public int probPick(int x, int y, Particle<int[]> p) {
+    public int probPick(int x, int y, Particle<double[]> p) {
         double k1 = CoreFactoryCreator.getFactory().getInstance().getK1();
         return (int) (Math.pow(
                 k1 / (k1 + itemAverageSimilarity(x, y, CoreFactoryCreator.getFactory().getInstance().getRadio(), p)), 2)
@@ -15,20 +15,20 @@ public class IrisFunction implements Function<int[]> {
     }
 
     @Override
-    public int probDeposit(int x, int y, Particle<int[]> p) {
+    public int probDeposit(int x, int y, Particle<double[]> p) {
         double fx = itemAverageSimilarity(y, x, CoreFactoryCreator.getFactory().getInstance().getRadio(), p);
         double k2 = CoreFactoryCreator.getFactory().getInstance().getK2();
         return (int) ((fx < k2 ? 2*fx : 1.0) * 100.0);
     }
 
     @Override
-    public double itemAverageSimilarity(int x, int y, int r, Particle<int[]> p) {
+    public double itemAverageSimilarity(int x, int y, int r, Particle<double[]> p) {
         System.out.println("Esta funcion no debe llamarse");
         return 0.0;
     }
 
     @Override
-    public double itemPerceivedFraction(int x, int y, int r, Particle<int[]> p) {
+    public double itemPerceivedFraction(int x, int y, int r, Particle<double[]> p) {
         int maxX = CoreFactoryCreator.getFactory().getInstance().getMaxX();
         int maxY = CoreFactoryCreator.getFactory().getInstance().getMaxY();
         double alpha = CoreFactoryCreator.getFactory().getInstance().getAlpha();
